+++ src/modules/teams/index.ts
@@ -65,7 +65,7 @@ export const teams = (
 
                 return {
                     ...team,
-                    fame: team.fame + getFameFromPlacement(place),
+                    fame: team.fame + (action.payload.isForFame ? getFameFromPlacement(place) : 0),
                     history: [...team.history, {
                         tournamentId: action.payload.tournamentId,
                         place,